feat(orga): moved tests and usecase
